+++ museum/src/components/Footer/Footer.css
@@ -1,7 +1,6 @@
 .footer {
   display: flex;
   background-color: #060403;
-  
   justify-content: center;
   color: #E8DBCB;
 }
