--- museum/src/components/Footer/FooterAddress.js
@@ -1,17 +1,0 @@
-import React from "react";
-
-
-const FooterAddress = () => {
-  return (
-    <div>
-      <h3>Coordonnées</h3>
-      <p>Museumstraat 1,</p>
-      <p>Amsterdam</p>
-      <br />
-      <h4>Horaires des visites:</h4>
-      <p>Tous les jours de 9h à 17h</p>
-    </div>
-  );
-};
-
-export default FooterAddress;
