+++ museum/src/components/Footer/FooterContent/FooterMap/FooterMapButton.js
@@ -1,13 +1,12 @@
 import React from "react";
-
 import FooterModalMap from "./FooterModalMap";
 
 const FooterMapButton = () => {
 
   return (
-    <div className="footerButton">
+    <>
       <FooterModalMap />
-    </div>
+    </>
   );
 };
