+++ museum/src/components/Footer/FooterContent/Footerlogo/FooterLogo.js
@@ -1,0 +1,12 @@
+import React from "react";
+import { Image } from "mui-image";
+
+const FooterLogo = () => {
+  return (
+    <>
+      <Image src="./logo-footer.png" alt="logo" width={{ xs: "25%" }} />
+    </>
+  );
+};
+
+export default FooterLogo;
