--- museum/src/components/Footer/FooterLogo.js
@@ -1,14 +1,0 @@
-import React from "react";
-import "./Footer.css";
-
-const FooterLogo = () => {
-    
-  return (
-    <div className="footerLogo">
-      <h3>WITH US</h3>
-      <img src="./logo-footer.png" alt="logo" />
-    </div>
-  );
-};
-
-export default FooterLogo;
